review: Refactoring SearchFilter에서 LetterButtonList 컴포넌트 분리 및 Click 이벤트 네이밍 변경  onClick => onLetterClick
